fix: price Long -> Double 대대적 타입 변경

+++ backend/funco-member/src/main/java/com/found_404/funcomember/feignClient/client/TradeServiceClient.java
@@ -19,7 +19,7 @@ public interface TradeServiceClient {
     CoinValuationResponse getCoinValuations(@RequestParam Long memberId);
 
     @GetMapping("/api/crypto")
-    Map<String, Long> getCryptoPrice(@RequestParam List<String> tickers);
+    Map<String, Double> getCryptoPrice(@RequestParam List<String> tickers);
 
     @GetMapping("/api/trade/holding/coins")
     List<HoldingCoinResponse> getHoldingCoinInfos(@RequestParam Long memberId);
